+++ index.html
@@ -3,10 +3,10 @@
 <head>
   <meta charset="utf-8">
   <meta name="viewport" content="width=device-width,initial-scale=1">
+  <link rel="stylesheet" htef="vendor/nouislider/nouislider.css">
+  <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
   <link rel="stylesheet" href="css/normalize.css">
   <link rel="stylesheet" href="css/style.css">
-  <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
-  <link rel="stylesheet" htef="vendor/nouislider/nouislider.css">
   <title>Кекстаграм</title>
 </head>
 
@@ -169,8 +169,8 @@
     </section>
   </template>
   <script src="vendor/pristine/pristine.min.js"></script>
-  <script src="js/main.js" type="module"></script>
   <script src="vendor/nouislider/nouislider.js"></script>
+  <script src="js/main.js" type="module"></script>
 
 </body>
 </html>
