+++ js/main.js
@@ -2,7 +2,7 @@ import {createThumbnails} from './create-thumbnails.js';
 import {createPictures} from './create-pictures.js';
 import {initOpenCloseForm} from './open-close-form.js';
 import {initValidatePristine} from './validate-pristine.js';
-import {pictureResize} from './size.js';
+import {initResizePicture} from './resize-picture.js';
 import {initPictureFilters} from './picture-filters.js';
 
 const pictures = createPictures();
@@ -10,5 +10,5 @@ const pictures = createPictures();
 createThumbnails(pictures);
 initValidatePristine();
 initOpenCloseForm();
-pictureResize();
+initResizePicture();
 initPictureFilters();
